Added move semantics to Stopwatch
Since Stopwatch had an explicit destructor (which writes
the timers in the log), the compiler was not generating
the move constructor and assignment.

Adding them will enable dealing with vectors of Stopwatches
without the need to resort to vectors of pointers of
Stopwatches.

I also made the Stopwatch explicitly not-copyable, since it does
not make sense to take a copy.