Changed the way -nan is converted to nan in fixzeros script so it works for tests involving grids as well as tests involving lepton
If your output contains nans you should use a format specifier like %8.4f and the fixzeros script will then work if you have a compiler that
outputs -nan.  If you use %f or %.7f then I think that nan is put at the start of the space.  If you use %8.4f then the nan will be put in the final
three characters of the format specifier