Fix in shape checking of cmd arguments.
Before this fix, the shape checking was done verifying
that the user is passing an array that, along each axis,
is at least as large as the requested one. This is not
very meaningful: actually all dimensions except the first
should be equal to the requested one.

For instance, passing 10 coordinates with a numpy array
with shape (10,4) makes no sense and should trigger an exception.