Revert to old syntax for CONTACT_MATRIX
In the old version of PLUMED when you created a contact matrix you did this

c1: CONTACT_MATRIX GROUP=1-100
PRINT ARG=c1

The print here would then access the matrix.  The new version replaced this with

c1: CONTACT_MATRIX GROUP=1-100
PRINT ARG=c1.w

as you can also do:

c1: CONTACT_MATRIX GROUP=1-100 COMPONENTS
PRINT ARG=c1.w

and have access to four matrices that contain the x, y and z components of the vectors connecting the various atoms.

This change reverts back to the first style when you do not use the COMPONENTS keyword so as to maintain back compatibility.
This changes the input for many (new) regtests, however, which is why this change seems so large