Fixed grids so that old style multicolvar still works.
In old multicolvar label of argument containing less than is aa.lessthan.
This is changed now to aa_lessthan as aa_lessthan is from a different action than
aa multicolvar.  I added to grid something to check if there is a value called aa.lessthan
in the grid file if it finds a request for aa_lessthan.