Removed special case for C arrays
Before this commit I was using a special case to pass arguments
by reference, to avoid decays of C arrays to pointer.

Now I found a way to do it using only the universal reference
implementation (T&&val), but using std::remove_reference to
properly detect C arrays.

This implementation is more portable to all compilers. It worked
with GCC 4.9, so I assume this is standard enough C++11 not to need
any special flag to switch it off