Improved LOAD action
This is a proof of concept that makes the LOAD action
working also when multiple Plumed objects exist, including
multithreading.

Note that the `plumed mklib` command is not threadsafe per se.
This means that thread safety is achieved only when loading directly
an already compiled shared library.

Also notice that so far only action register has been fixed. This might
be sufficient or not, depending what loaded libraries are adding.