added mutex on call to mklib
This is making sure that two calls to `plumed mklib` with
identical arguments are not executed simultaneously.

Notice that, even if they were, there would be no error, just
overload due to multiple redundante compilations happening.

Also notice that we can only check threads belonging to the same
process in this way. Multiple MPI replicas (with an intercomm)
should still lead to harmless duplicated work.