Made register generic
I made a new generic class to be used for registers.
It is used in both actionRegister and cltoolRegister.

In addition, the `--load` command line option properly works
in a thread safe context (i.e., if multiple command line actions
are run on multiple threads).

Notice that action loaded with --load are not seen by the plumed
engine, but only to the command line interpreter. So this is just
to load separate command line tools.