Added lock when accessing the register for reading.
I didn't observe any deadlock heuristically. However,
I am sure that reading the std::map with the register should
not be done while the maps are modified. So, for thread safety,
I added a lock that ensures that, when a thread is modifying the map,
no other thread can modify it or read it.