Removed Doxygen characters from first line of documentation
In the nest and tutorials the first line of the documentation for actions appears in the tooltip
that appears when you hover over the action name.  This line of text should thus provide a brief summary of
what the action does.  To make this renderable this description should not include anything that needs to be
rendered by Doxygen (i.e. equations and ref commands).

This commit adds a check when the json syntax file is rendered to ensure folks do not have backslash characters in
that first line of syntax. I think this should capture all the doxygen commands that people might use in these lines.

It also removes the few places where people have included equations and so on in that first line of the documentation