Added information about what is contained in the values that are calculated by each action
When we introduced docstrings to describe the values stored in each of the components for an
action we never added the docstrings for the value.  In other words, if you have a command like this:

v: ACTION ...

There is no documentation for the Value with label v.  There is only documentation for v.component.
This commit resolves this problem and adds documentation for all values.