Fixes to make inputs from tutorial pages work with master version of code.
This fix includes addition of code to read the NORMALIZATION keyword in AVERAGE.  The user is shown a warning if they use this keyword and recommended to use the new (Simpler) syntax.

I also added the functionaility to use the atom selection tools from ActionAtomistic to specify the ATOM_INDICES in DUMPPDB.  In doing this I made interpretAtomList a static function so
I can reuse it in actions that do not inherit from ActionAtomistic.