Re-engineered changes to interpretAtomList from last but one commit
I have added a function that is a non-static member to ActionAtomistic called interpretAtomList.
This non static version means that you can use interpretAtomList in the old way (the function
is a call to the static function that I implemented).  You can thus continue to use interpretAtomList
in the old way or, if you want to call the method from a class that doesn't inherit from ActionAtomistic
you can call the static version of the function