Modified code for creating the dictionary of values that are computed by shortcuts.
The names of the actions that have descriptions in the keyword object for a shortcut are now stored at the same
time the lines of the shortcut are stored (i.e. when the shortcut actions are created).  When driver asks for
documentation for these components it thus gets the relevant actions from the shortcut and then looks up the corresponding
documentation strings