Reduced the lengths of the long lines in shortcuts that are output
If the lines of shortcut are very long then the plumed nest runs very slowly as you have to do
a lot of i/o.  I have thus written some code here to reduce the lines that are particularly long
in a way that does not remove the understanding from the resulting shortcut that is output.