Fixed derivatives for matrix vector multiplication when matrix is sparse and stored
This commit fixes a bug with the derivatives for matrix-vector multiplication that occurs when you store the derivatives
and compute the forces in the back-propegation (apply) step instead of calculating the derivatives in the forward (calculate) loop by using
the chain.  The problems occur when the matrix that is being used is sparse.

The feature that is fixed here is not currently used as in all the cases it could be used the derivatives are calculated during the calculate step.
I think it is better to fix it here, however, as it may be used in the future.

Notice that chnages were required to KDE because this does apply forces to a sparse matrix.  In the old version of the code when applying forces in KDE
the matrix was assumed to be full even if it was sparse.  The changes to KDE are thus using the sparsity more effectively when doing the back propegation
for the forces.