Optimisation of MatrixVectorProduct to improve performance when multiplying matrix by vector of ones
If you are multiplying the matrix by a vector of ones you don't need to do any multiplication.  You can just sum each row of the matrix.
This is now used to calculate this type of matrix-vector product.  The change speeds up the code slightly in an operation that is done reasonably often.
Changes to regtests are due to the derivatives on the constant vector of ones no longer being stored or used.