Replaced checking of derivatives for many regtests with testing of forces
In the old version of the code if you have a input like this:

ones: ONES SIZE=100
cmap: CONTACT_MAP GROUP=1-100
coord: MATRIX_TIMES_VECTOR ARG=cmap,ones
sum: SUM ARG=coord PERIODIC=NO
DUMPDERIVATIVES ARG=sum

The derivatives printed out would be the derivatives of the sum with respect to the atom positions
that were input to the action labelled cmap.  This was the chain facilitated.

When the chain is not present sum outputs derivatives with respect ot the input - the list of coordination
numbers output by coord.  In other words, the derivative is a vector of 100 ones.  Outputting this is not very
useful so to ensure that derivatives throughout the input are the same it is better to output the forces that were
obtained using a biasvalue acting upon sum.  This test is then equivalent to the old test on the outputted derivatives