Fixed calculation for rt-lq6-vol
I have removed the test on the forces for this regtest as the code with the new derivatives
performs this calculation very slowly because I am no longer able to work out the subset of
atoms on which to calculate the q6 parameter.  Also I think this test is not very robust when the calculation
is done in different ways.

I am pretty certain that the forces for this type of calculation are calculated correctly as there are other tests
that test forces that work in similar situations.  However, some careful benchmarking of this type of calculation
needs to be peformed before this branch is merged