FunctionOfMatrix and QuaternionBondProductMatrix are now parallelized over matrix elements rather than over the rows of the matrix
These clases now inherit from ActionWithVector rather than ActionWithMatrix

Changes to regtests revert changes that were made in an earlier commit when I started  reworking the
derivatives.  In other words, the forces in these actions are back to the values that we had when
we were using the chains derivatives.  I had introduced an error in an earlier commit