Use process group signal handling to reach past the shell
Spawn processes is their own process group and then signal the
process group intstead of just the process we spawned to ensure
the signals also reach any subprocesses started by the process we
spawned (which is always the case as we are wrapping all our
processes spawning with "/bin/sh -c ...").

This make sending SIGINT work again, so revert back to that instead
of SIGKILL. The reason SIGINT wasn't working was the shell logic
is that this is expected to be sent to the process group, so it
doesn't need to pass it along or anything. See the bash signals
documentation page for more details

https://www.gnu.org/software/bash/manual/html_node/Signals.html