Fixed some issues with regtests
This fixes some issues with the newer regtests on secondary structure variables.  I have removed the test
secondarystructure/rt-coverage-parabetarmsd-parabeta-intra as it is identical to rt33 with the exception of
adding a secondary structure variable that is exactly zero because you are not considering any segments.  I have
also added checks in the constructors for AntibetaRMSD and ParabetaRMSD to ensure that the user has specified at least
one segment to compare with the ideal structure.

For the other test that had to be fixed I specified the specific force values that I wanted to print to the output file and
all problems were resolved.