Added new method for links to switching function documentation
This commit addresses something that has bothered me for a while; namely, how we should
provide the links to the parts of the manual on swiching function when you encounter a keyword
like switch.  I would ideally like to provide the link to this page both in the tooltip that is
generated for the annotated inputs and in the tables that contain the syntax.

This commit addresses this problem.  You can now use the command keys.linkActionInDocs when you
register the keyword in registerKeywords.  Links to the manual pages for relevant actions will then
be included in the manual and in tooltips.

As part of this I have also moved the documentation for switching function to the action LESS_THAN
and the documentation for HistogramBead to BETWEEN as this is where these classes are used in their
most basic form.