Fix for problems with regtests
The change to the regtest here occurs because we now call calculate in the generate mask function.
If the landmarks are selected randomly this changes the set of landmarks that are selected because you
have an additional call to calculate.  In other words, the regtest values have changed because different
random numbers are being generated.  The code functionality has not changed