Reduced precision of numbers in output files for two regtests for ves module that are failing on github (#1231)
To make this work I added an option GRID_FMT that sets the precision of the grids that are output by the
VES_LINEAR_EXPANSION action.  The same format is then used for all the output files that
are output by this command.  This is why there are more changes to the regtests than just the single file
that is currently failing.

Co-authored-by: Gareth Aneurin Tribello <garethtribello@Gareths-MacBook-Pro.local>