Prior to this small change rt-smap1 was running sketch-map with no data (as it runs on step 1).  Now that doesn't happen.
Sketch-map was running no data as we are running sketch-map with a stride of 50 and the first step is step 0.  Hence, the
sketch-map calculation is run on the first step.  This is harmless unless you are running without optimisation.  If you run
without optimisation the calculation just hangs because it gets stuck trying to diagonalize a matrix in which every element is nan
for MDS