Evaluating KDE using the ParallelTaskManager
N.B. I changed the way that the normalization factor for gaussian histograms is computed.
The equation is equivalent but there are some numerical precision differences in the way
it is computed now and the way we used to compute it.  I think that the new way ensures that
the log more clearly indicates what is going on so I would like to keep it.  The numerical
differences are the reasons why there are a number of changes in the regtests.

N.B2. changes in input for contour/rt-wcsurface2/plumed.dat was made because the old code
had a counterintuitive behaviour.  I have fixed this behaviour but it means that the values
in the rt-wcsurface2 regtest change.  I can reproduce the old values by making the changes that
I made in the plumed.dat file for this regtest