Fixed derivatives with respect to grid point coordinates in spherical kernels
The derivatives of the grid points if you do a spherical KDE were computed wrongly,
which is why I have changed this regtest.  The derivatives that were output were computed
using the derivatives with respect to the positions of the kernels rather than the derivatives
with respect to the position of the grid points.  This is why the regtest changes.  It makes
no difference to any key functionality as these derivatives are not used for anything within the
code.  I also don't know how to interpolate a function on a sphere.