Improved implementation of pairentropies CV to make calculation more parallelised
Also rechecked the numbers output by the versions of the PAIRENTROPIES and PAIRENTROPY command
with the numbers output by Pablo's implementation of these actions. In doing this check
I discovered that PAIRENTROPIES was not implemented correctly in PLUMED and that the values
output by this action were wrong.  Changes to the reference values of this quantity in the regtest
rt-pairentropies were thus necessary.